Copy custom coverage editing to full-page view
