+++ src/components/coverage-statement-fields/coverage-statement-fields.css
@@ -1,0 +1,3 @@
+.coverage-statement-textarea {
+  max-width: 50em;
+}
