+++ src/components/coverage-statement-fields/coverage-statement-fields.js
@@ -2,12 +2,14 @@ import React, { Component } from 'react';
 import { Field } from 'redux-form';
 
 import TextArea from '@folio/stripes-components/lib/TextArea';
+import styles from './coverage-statement-fields.css';
 
 export default class CoverageStatementFields extends Component {
   render() {
     return (
       <div
         data-test-eholdings-coverage-statement-textarea
+        className={styles['coverage-statement-textarea']}
       >
         <Field
           name="coverageStatement"
