+++ src/components/coverage-statement-fields/index.js
@@ -1,0 +1,1 @@
+export { default, validate } from './coverage-statement-fields';
