+++ src/components/customer-resource-coverage-fields/customer-resource-coverage-fields.css
@@ -1,0 +1,31 @@
+@import '@folio/stripes-components/lib/variables';
+
+.coverage-fields {
+  max-width: 50em;
+}
+
+.coverage-fields-date-range-rows {
+  list-style-type: none;
+  padding: 0;
+}
+
+.coverage-fields-date-range-row {
+  display: flex;
+  align-items: flex-start;
+}
+
+.coverage-fields-datepicker {
+  align-self: flex-start;
+  flex: 1 1 200px;
+  margin-right: 1em;
+}
+
+.coverage-fields-date-range-clear-row {
+  flex: 0 0 3em;
+  margin-top: 2.25em;
+  text-align: center;
+
+  @media (--mediumUp) {
+    margin-top: 1.5em;
+  }
+}
