+++ src/components/customer-resource-coverage-fields/index.js
@@ -1,0 +1,1 @@
+export { default, validate } from './customer-resource-coverage-fields';
