+++ src/components/customer-resource-coverage/customer-resource-coverage.css
@@ -12,40 +12,6 @@
   align-items: center;
 }
 
-.coverage-form-date-range-rows {
-  list-style-type: none;
-  padding: 0;
-}
-
-.coverage-form-date-range-row {
-  display: flex;
-  align-items: flex-start;
-}
-
-.coverage-form-datepicker {
-  align-self: flex-start;
-  flex: 1 1 200px;
-  margin-right: 1em;
-}
-
-.coverage-form-date-range-clear-row {
-  flex: 0 0 3em;
-  margin-top: 2.25em;
-  text-align: center;
-
-  @media (--mediumUp) {
-    margin-top: 1.5em;
-  }
-}
-
-.coverage-form-add-row-button button {
-  width: 100%;
-
-  @media (--mediumUp) {
-    width: auto;
-  }
-}
-
 .coverage-form-action-buttons {
   border-top: 1px solid var(--minor-divider-color);
   display: flex;
