+++ tests/pages/customer-resource-edit.js
@@ -1,6 +1,8 @@
 import {
+  action,
   blurrable,
   clickable,
+  collection,
   fillable,
   isPresent,
   page,
@@ -9,6 +11,7 @@ import {
   value
 } from '@bigtest/interaction';
 import { hasClassBeginningWith } from './helpers';
+import Datepicker from './datepicker';
 
 @page class CustomerResourceEditNavigationModal {}
 
@@ -16,9 +19,21 @@ import { hasClassBeginningWith } from './helpers';
   navigationModal = new CustomerResourceEditNavigationModal('#navigation-modal');
 
   clickCancel = clickable('[data-test-eholdings-customer-resource-cancel-button] button');
-  clickSave = clickable('[data-test-eholdings-customer-resource-cancel-button] button');
-  isSaveDisabled = property('disabled', '[data-test-eholdings-customer-resource-cancel-button] button');
+  clickSave = clickable('[data-test-eholdings-customer-resource-save-button] button');
+  isSaveDisabled = property('disabled', '[data-test-eholdings-customer-resource-save-button] button');
   hasErrors = isPresent('[data-test-eholdings-details-view-error="resource"]');
+
+  clickAddRowButton = clickable('[data-test-eholdings-coverage-fields-add-row-button] button');
+
+  dateRangeRowList = collection('[data-test-eholdings-coverage-fields-date-range-row]', {
+    beginDate: new Datepicker('[data-test-eholdings-coverage-fields-date-range-begin]'),
+    endDate: new Datepicker('[data-test-eholdings-coverage-fields-date-range-end]'),
+    clickRemoveRowButton: clickable('[data-test-eholdings-coverage-fields-remove-row-button] button'),
+    fillDates(beginDate, endDate) {
+      return this.beginDate.fillAndBlur(beginDate)
+        .append(this.endDate.fillAndBlur(endDate));
+    }
+  });
 
   coverageStatement = value('[data-test-eholdings-coverage-statement-textarea] textarea');
   fillCoverageStatement = fillable('[data-test-eholdings-coverage-statement-textarea] textarea');
@@ -26,11 +41,11 @@ import { hasClassBeginningWith } from './helpers';
   coverageStatementHasError = hasClassBeginningWith('feedbackError--', '[data-test-eholdings-coverage-statement-textarea] textarea');
   validationErrorOnCoverageStatement = text('[data-test-eholdings-coverage-statement-textarea] [class^="feedbackError--"]');
 
-  inputCoverageStatement(statement) {
+  inputCoverageStatement = action(function (statement) {
     return this
       .fillCoverageStatement(statement)
       .blurCoverageStatement();
-  }
+  })
 }
 
 export default new CustomerResourceEditPage('[data-test-eholdings-details-view="resource"]');
